Actualizacion.
Se crean la funciones para la consulta de datos a la base de datos para los reportes.

+++ codigo/GestionHotelera/Models/FacturasYReservasModel/FacturasModel.cs
@@ -13,7 +13,7 @@
 
         public int CantidadPersonas { get; set; }
 
-        public int PrecioTotal { get; set; }
+        public double PrecioTotal { get; set; }
 
         public int NumeroHabitacion { get; set; }
